// app/types/alumni.ts
export interface Alumni {
  full_name: string;
  avatar: string;
  email: string;
  current_company: string;
  current_company_position: string;
  education_1: string;
  education_degree_1: string;
  education_start_1: string;
  education_end_1: string;
  education_fos_1: string;
  education_2: string;
  education_degree_2: string;
  education_start_2: string;
  education_end_2: string;
  education_fos_2: string;
  skills: string;
  location_name: string;
  headline: string;
  profile_url: string;
}